LayoutTests/fast/forms/file/entries-api/pages-jpeg-open-panel-expected.txt

OPEN FILE PANEL
Tests that a Pages file, chosen via the file picker which accepts Pages files and JPEG images, has a non-zero file size. To test manually, click on the file input and select a Pages file.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fileList.length is 1
PASS file.size is 89650
PASS successfullyParsed is true

TEST COMPLETE
